fastapi_app/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
langchain==0.1.0
langchain-community==0.0.13
langchain-google-genai==0.0.5
google-generativeai==0.3.2
pydantic-settings==2.1.0
chromadb==0.4.22
transformers==4.52.4
huggingface_hub==0.31.1
pysqlite3-binary==0.5.4
pandas==2.2.3
prometheus_client
beautifulsoup4
selenium
boto3